Pass the "Human Folders" Option Through the `Library` Class
To remove the need for the `Game` class to parse `sys.argv`, we'll store
the option in the `Library` class and pass it to `Game` as needed.

The `ArgumentParser.add_argument()` call has been updated to show up as an
'on/off' argument.

Also fix the naming convention. According to Python's PEP 8 Style Guide,
function and variable names should be lowercase with words separated by
underscores (aka snake_case). The rest of this project seems to follow PEP
8, so `humanFolders` should be `human_folders`.