Fix broken `--download-game` parameter
In commit c652516eec390606d81350dd6f6a1c32e52cf747 a check was added to
determine if a game was not purchased via the `uploads/` endpoint.
Unfortunately that endpoint replaced the call to the game data endpoint
that we need to download the game. This adds the game data endpoint call
back in.